+++ homework11/KMP/textSearch.cpp
@@ -1,3 +1,4 @@
+#include "textSearch.h"
 #include <string>
 #include <vector>
 #include <fstream>
@@ -7,7 +8,7 @@ std::vector<int> prefix(const std::string &source)
 	std::vector<int> prefixes;
 	prefixes.push_back(0);
 	
-	for (int i = 1; i < source.length(); ++i)
+	for (size_t i = 1; i < source.length(); ++i)
 	{
 		int longestPrefix = prefixes[i - 1];
